backend/agents/sentiment_analyst.py
"""Sentiment Analyst Agent — Analyzes news sentiment."""

from crewai import Agent

SA_ROLE = "Sentiment Analyst"
SA_GOAL = (
    "Analyze news sentiment for candidate stocks. For each stock, evaluate the last 5 headlines "
    "and score overall sentiment from -1 (very negative) to +1 (very positive). "
    "Flag stocks with strong positive or negative news catalysts."
)
SA_BACKSTORY = (
    "You are a market sentiment specialist who reads financial news and social media "
    "to gauge market sentiment around stocks. You can quickly assess whether news is "
    "positive, negative, or neutral for a stock's price action. You provide sentiment "
    "scores and identify the most impactful headline for each stock."
)


def create_sentiment_analyst(llm) -> Agent:
    return Agent(
        role=SA_ROLE,
        goal=SA_GOAL,
        backstory=SA_BACKSTORY,
        verbose=True,
        llm=llm,
        allow_delegation=False,
    )
