backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
sse-starlette==2.1.0
groq==0.11.0
yfinance>=0.2.40
pandas==2.2.3
numpy==1.26.4
ta==0.11.0
python-dotenv==1.0.1
httpx==0.27.2
aiohttp==3.10.5
cachetools==5.5.0
pydantic==2.9.2
# Optional — CrewAI (requires Linux/Mac; uvloop does not build on Windows)
# pip install crewai==0.80.0 crewai-tools==0.14.0
